# go/model_message.go
/*
 * Secret Message Exchange Service
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Message struct {

	Message string `json:"message,omitempty"`
	// Time to live in seconds
	Ttl int32 `json:"ttl,omitempty"`
	// Maximum number of reads allowed
	MaxReads int32 `json:"max_reads,omitempty"`
}